requirements.txt
## Install all
# Run `pip install -r requirements.txt` to install all modules.

# pip install websockets
websockets=15.0.1

# pip install twitchio[starlette]
twitchio[starlette]=3.0.1

aiohttp==3.9.5
python-dotenv==1.0.1
requests==2.31.0
